2023 Day 19 Part 2
Okay, now we're doing combinatorics! It looks scary at first, but fundamentally it's just range splitting based on value which pass or fail the test. I initially was going to actually deal with range objects, but then I realized that the list of numbers is probably small enough that I can just iterate through all the options for each rule test. It's less efficient, but way less to code! (And it does work!)

Rank 110, just missed! Definitely would have leaderboarded were it not for part 1...