2023 Fix Day 16
I realized that I was a bit overzealous in optimizing the day 16 code. I assumed that hitting a splitter on the side (causing a split) was required for a loop, but today I realized that it's possible to construct a loop that goes through a splitter but only causes a split when *entering* the loop, not on each cycle. As such, to avoid infinite loops shot_ray (formerly shoot_to_split) needs to return whenever it hits a splitter, regardless of the angle.

Example inputs which caused infinite loops on the previous old code:
.\.
/-\
\./

\/\
\|.
.\/