2023 Day 20 Part 2
Hoo boy! I was expecting something tricky in part 2 given this graph setup, and I was not disappointed or surprised.

First thing I tried was a brute force. I figured it wouldn't work, but it wasn't hard to cook up and I could leave it running in the background as I worked through a real solution.

With that out of the way, an inspection of the input reveals that rx has a single conjunction module as an input, and that conjunction module has 4 conjunction modules as its inputs. Suspiciously, broadcaster sends to 4 modulse itself! After a bit of time (probably too much time) quickly sanity checking I guessed that there are 4 distinct sub-graphs in the network, each with their own period.

With these subgraphs having their own period all we have to do is detect each subgraph's period and take the LCM of the periods to get the answer. I didn't fully vet that the network cleanly splits in that manner (that would take *way* too much time to code) but in the end the answer was accepted so it definitely seems to be the case! I'll have to verify this more properly when refactoring this solution.

I got rank 18! Yes! Glad I spotted the pattern/trick relatively quickly on this one.