2023 Day 21 Part 2
Wow. I normally wouldn't commit my solution *quite* this raw, but I spent ages trying a myriad of things and I'm going to commit the whole mess for posterity.

Where to even begin. At first I thought it might be good to handle it like parallel universes in Super Mario 64, counting how many "instances" are on each tile in the main grid. This doesn't work (at least not how I did it) because:
1) The iteration count we need to do is insane
2) This won't give the right answer anyway because we need to deduplicate stepping on the same coordinate, while also counting two coordinates from different universes that map to the same home coordinate.
Maybe there's a clever way to deal with both #1 and #2, but I never figured it out.

After that I tried a more direct simulation just to see how bad it was. Pretty bad, as I anticipated.

Then I tried to analyize how this monster grows. I had missteps along the way but had been making progress with a partial simulation (to 65+131*3 steps) and data extracted to count how many squares were in various grid universe "types" at the end of the whole simulation. I kept getting the math wrong though.

I then spent just over 3 hours debugging and trying a fully analytical approach until I got the right answer, which was a huge pain. (It didn't help that I was doing it all in my head instead of sketching anything. I finally cracked it when I sketched out the growth pattern over 3 cycles of 131 steps!)

As part of this I kept dumping out partial grids from both my simulation and my calculation branches to try to figure out what was going wrong. Now that I know the growth pattern and characteristics I maybe could solve this just with simulating for 65+131*2 steps and extracting info from that simulation, but that is a problem for some other time. I need sleep.

Rank 2145. I'm not sure if I was just slow in analyzing this or if I missed a huge trick which made this simpler...