2023 Day 22 Part 1
For a moment when reading the problem I thought we might need to do tetris, but thankfully no! We do need to look out for intersections just like tetris, but no row disintegrations.

One important observation: If we sort the initial list of bricks based on their lowest z point then it is guaranteed that every brick must be supported by the ground or by a brick that came earlier in the list.

Fundamentally this is pretty simple with sets for the points in a brick, as well as a set for all the currently settled positions. The 1456 bricks are only comprised of 4211 bricks so while a smarter solution will run faster, a dumb O(n^2) algorithm works fine for seeing if any bricks are safe to disintegrate, just test all subsequent bricks and see if they fall!

I did have a dumb mistake though: When testing to see if bricks fell after disintigration I forgot to remove them from the settled positions set too. This meant that I got false positives for supported bricks took 12 minutes to debug...

Only got rank 189, so that bug cost me a lot. In fact, my original solution submission time would have been around rank 25 had I not had that bug!