2023 Day 22 Part 2
Okay, so part 2 is basically the same as part 1 except that we count how many bricks fall! A slight tweak to the dumb O(n^2) algorithm and it worked fine, the main thing to be concerned with is that settled_positions needs to be updated to count how many bricks fall so we need to store a copy of the original set to restore at the end. It's definitely slower than part 1, but we're still only talking about 1456 bricks so it's faster than writing something smarter!

Okay good, I managed to jump to 93 and leaderboard (barely). But man, I'd have been around rank 10 without my part 1 bug...