2023 Rewrite Day 22
All we really need is the support graph per brick and we can solve parts 1 and 2 very directly with graph relation queries. This saves a ton of time spent in simulating brick falls all the time.

Also, the initial brick settling logic can be optimized a good bit. Specifically, if the highest z point is tracked per x,y coordinate then it is possible to directly compute how low the brick will go before settling.

All of this cuts the time spent by the code into a tiny fraction of what it used to take, something on the order of 0.05 seconds total for both parts combined.