2023 Day 23 Part 1
Okay, this one threw me for a loop! I was initially thinking of compressing the grid into a graph between points of interest, namely the start point, end point, and slopes. However, this doesn't fully protect against stepping on the same location twice since slopes might point to the same intersection! (In retrospect I should have compressed this into a graph connecting the intersections...)

I even thought I might be being clever by finding the dijkstra length between the start and end with negative distances, but that didn't work. For one, it hits the above issue, and for another, I don't think Dijkstra is guaranteed to find the best path when negative path weights are in play. (I did just realize how I can make it work while typing this out, something to do in a cleanup pass...)

Anyway, I turned to a BFS instead which also checks for double-stepping. This is not fast, but it works. However, I accidentally had the negative Dijkstra line left after my new code for a bit which was blowing away my answer, took me some time to realize that's why I was still getting the wrong answer...

Rank 805, ouch. Not good!