2023 Day 23 Part 2
And part 2 eliminates the slopes which explodes the search space further! Here's where I realized that I needed to compress to a graph connecting intersections. I whipped that up, though I did have a bug for a while where I accidentally allowed stepping into forest tiles on the *first* step out of an intersection. Other than that change the approach is about the same as part 1 was, and likewise slow.

In fact, it was slow enough that I ended up letting it run in the background and work on rewriting part 1 with a graph connecting intersections. This turned out to be good as it helped me spot some of the issues I had in the code (such as allowing stepping into forest tiles) which was making part 2 take even longer, and it would have caused it to get the wrong answer! But I didn't finish the rewrite before the final background solve finished so that will be in its own commit.

Rank 214. Better, but still bad for today. But at least I have it solved.