2023 Rewrite Day 23
Parts 1 and 2 are fundamentally the same so long as the input is preprocessed in part 2 to eliminate slopes. This can be exploited to extract a single solve function.

Furthermore, finding the longest path via recursion is a good bit faster than the iterative method I had coded up. Part 2 still takes about 35 seconds though, I feel like there has to be some other trick I'm still missing to make this run faster but that's a problem for another time.