2023 Day 24 Part 1
Man, this threw me for a loop.

Initially I thought it was just asking which hailstones enter the test area. I don't know how I misread that, but that's definitely a different problem.

Then I thought it was asking if two hailstones would be in the test area at the same time. Got that coded up but again it was wrong.

Then I finally realized it was asking where they intersect, but I've not manually done line-line intersections in a while so I goofed. At one point I tried to send z3 to solve it but it took a while, and it didn't even give me an answer on my data because I forgot to reset the test area when switching from the sample input. Thankfully I was already algebraically solving the system of equations by hand and had a better method nearly coded up by then so I didn't have to wait for z3 yet again.

Rank 1027. Ouch. I clearly need a "Where do these intersect" utility in my graphics library...