2023 Day 24 Part 2
This also threw me for a loop, and it really shouldn't have thrown me for as much of a loop!

Fundamentally we have (for each hailstone) these 3 equations:
x + t * xv = p.x + t * v.x
y + t * yv = p.y + t * v.y
z + t * zv = p.z + t * v.z
Where x, y, z is the initial positions of the stone, t is the time of collision, and xv, yv, zv is the velocity of the stone.

Initially I tried sending those equations for *all* the hailstones at z3. It could do it for the example but it choked on my real input, sadly. I left that running in the background and tried some other things.
1) I misread the example and thought that it implied that at t=1 and t=2 there will always be a collision, so I tried iterating through pairs of hailstones to extrapolate the stone's position and velocity, but this never worked for the example. (The first two intersections are t=1 and t=3!)
2) I tried to solve the systems of equations for 3 hailstones and the stone by hand. This gets super messy...
3) I started to code up a 3-deep nested loop to have z3 try to solve the system of equations for 3 hailstones for me.

However, while working on #3 I realized that we're guaranteed that *any* three hailstones are hit and are solvable. As such, the nested loop is dumb! Given this, I tried setting up z3 to just solve the system of equations for the first 3 hailstones, but it choked on that too. Thankfully I remembered about sympy and after verifying its API, I gave it a shot. It pretty much immediately gave me the answer, finally!

I clearly don't do this sort of work enough, this was way more painful than it should have been.

Rank 502. Oof, wish I'd have thought of sympy sooner...