2023 Day 25 Part 1 (There is no Part 2)
I have to say, this leaves a very bad taste in my mouth. Apparently the only reasonable way to solve this is by using an external library which seems very much out of the spirit of Advent of Code. Not only that, but this is an NP-Complete problem, and seems entirely unsuited for the last day.

Rank 2576 (2168 for part 2). Whatever. Unless I'm missing something this is just a bad problem.