2024 Day 2 Part 2
More server issues, it took me multiple attempts to get my initial answer submitted! (But I also goofed...)

Anyway, dropping a number to check if a line is "safe" sounds easy and is...but I accidentally double-counted a number of lines that were safe to begin with and submitted an inflated answer. Lost ~3.5 minutes to that dumb mistake...

Rank 627. Missed the leaderboard by 4 minutes, 20 seconds. A big chunk was my dumb bug, but the server issues compounded that further. :(