2024 Day 5 Part 1
I was thinking we might have a graph problem today actually! This is an interesting sort of graph being a bunch of rules dictating ordering when two items show up in a list. That doesn't inherently guarantee that you have an acyclic graph in fact! Anyway, checking validity is pretty easy, albeit the Big-O complexity is trash.

Rank 212. Yeah, I did take a bit extra time waffling on how to parse the input. Missed the leaderboard by 1:28.