2024 Day 5 Part 2
Okay, I definitely dropped the ball speed-wise here. I juggled a number of ideas on how to properly reorder the bad lists before settling on a proper graph-like one, then took an irksome amount of time to write (and debug!) it! I was comparing lists with tuples which always failed, resulting in me not reordering the list at all. It's a good thing I added a check to verify that the reorder was successful, but this took an annoying amount of time to debug.

And in retrospect I realized I could have done a simpler, more brute-force approach by checking all positions for each addition and accepting the one that is allowed. Way less thought and faster to implement, even if the runtime is marginally worse...

All the way down to rank 753...missed the leaderboard by 10:11. Oof.