2024 Day 6 Part 2
Oof, and a lot fell apart here. My knee-jerk thought was to do a "smart" brute force because checking the whole grid felt too expensive. I really should have checked the size of the grid though as it turns out that my optimization would only have cut out 2/3rds of the search space so it was not worth the implementation time and complication...

Anyway, the thought was to only check locations where the guard actually walks (since that's where there could be loops) and then look for a loop when an obstacle is there. That's great, but I had 3 bugs:
1) Because I wasn't sure how long it would take I added progress reporting code. However, I accidentally only checked candidate positions when I reported progress which was every 100th candidate! Obviously this misses a lot...
2) Initially my loop detection code used a seen set which had the start position/vector inserted, implicitly accepting everything! Again, silly mistake.
3) Worst of all, even if that weren't there my code wasn't *actually* inserting an obstacle in the candidate location and still treated it as a free space!

Quite a dumpster fire, compounded by decisions that just complicated the solution. When I went back to a dumb brute force it was much quicker to implement and while not fast to run, it did give the right answer.

(It's also worth noting that I tried to implement another approach in the middle too, but it didn't pan out and isn't worth discussing. Just another waste of time.)

Rank 786. Missed the leaderboard by just under 12 minutes, but my first (bad) answer was within the leaderboard range so without my bugs I finally would have had points. Oof!