2024 Refactor and optimize Day 7
Both parts are essentially the same, with one extra operator allowed in part 2. As such a combined solve method is pretty easy to write!

Additionally, abort checking if an equation is fixable when the current value exceeds the target (all operations grow numbers so we've failed by that point) and implement the concat operation a little more directly to optimize part 2. Part 2 still takes about 1.65 seconds but this still is a good speedup.