2024 Further optimize Day 7
Verulean314 on the reddit had a very clever idea: Process the numbers in reverse! This excludes prunes down the search space much faster since multiplication is only possible if the target is a multiple of n, and concatenation is only possible if the target ends with the same digits. This runs super fast now, essentially instant!