2024 Day 8 Part 1
This was...confusing. Saying that an antinode occurs only when "one of the antennas is twice as far away as the other" is a pretty kludgy way to say that it is in line with the antenna and the same distance as the other antenna. (To the point that I only understood what that sentence was saying as I was writing this comment!) Took me a while looking at the examples to guess what the problem was, and even then it was just a guess that I didn't feel firm on.

On top of that, it had seemed that antinodes on top of antennas were excluded, but that's not true and lost me over a minute (1 minute from the timeout and some extra from extra implementation to exclude these overlaps)!

Rank 268. I definitely could have been faster had I understood the assignment quicker, not sure if I'd have leaderboarded since I missed by 2:16 but I'd have been way closer.