2024 Rewrite Day 8
Parts 1 and 2 are very similar, except for slightly different antinode generation logic. As such, it's possible to implement a common solve method which just takes in a function to generate antinodes, deduplicating much of the logic between the parts.

Additionally, ensure that truly all antinodes are generated in part 2, even if the slope is not "reduced". The problem input is kind in this regard but it is not strictly guaranteed!