2024 Day 9 Part 2
Oh hey, now we're moving the whole files to reduce fragmentation. I was anticipating this in part 1 (and it's part of what tripped me up!) This implementation is terrible, there should *not* be a quadratic search to find free blocks but I didn't want to spend the time implementing something better.

Rank 176. At least I jumped up, but still, pain. Would have been an easy leaderboard had I understood part 1 in a reasonable period of time...