2024 Rewrite Day 9
Keeping track of full file blocks makes part 2 much faster and simpler. Moreover, it allows for implementing nthistle's idea of tracking free blocks in heaps to more efficiently find the best free block without traversing a ton of irrelevant small blocks at the front of the disk. All told this gets the part 2 runtime down to ~0.04 seconds.

Additionally, part 1 can use the same implementation so long as the files are represented as multiple 1-block files. (Idea taken from Jonathan Paulson's implementation.

Finally, some extra care is needed to properly merge free blocks when files of length 0 appear in the disk map. While not present in the input, this would throw the solver for a loop with inputs such as 13030404509 (which should have an answer of 465).