2024 Rewrite Day 11
Parts 1 and 2 can use the same approach since the cycle count is all that differs so it's cleaner to have a single solve function. Also, instead of recursion and memoization this can be solved pretty easily with iteration and collections.Counter instead. This makes things a little cleaner and reduces the number of map keys to deal with, providing better runtime guarantees.

This does still bog down with large cycle counts, but I did observe that eventually the stone numbers seem to stabilize into a given set of numbers. As such, the relation can probably be encoded into a matrix and matrix exponentiation used to more quickly compute the final answer, I may look into that next.