2024 Day 12 Part 2
So this sort of fell apart. Figuring out which edges were part of the same side took me way too long to get working.

Initially I was thinking of tracking edges as "halfway" between the in and out sides and that sort of works, but if you're not careful you can "tunnel through" the region when regions have offshoots with a width of 1.

Then I started working on trying to find neighbor edge cells that were the same "direction" from the region. This does work, but I had a nasty little issue where I was "jumping" across gaps because some cells were part of the perimeter but were reached from the opposite direction! That took quite a while to nail down actually, something like 12 minutes! When I finally figured it out it was pretty easy to correct, but ouch.

Rank 710. What a mess of a day, and I know my solution is incredibly inelegant...