2024 Rewrite Day 12
The part 1 and 2 logic become much simpler with a unified Region class encapsulating some of the logic. This also makes it possible to write a single solve function taking a region scoring function making part 1 and 2 very short.

Most importantly though, the side counting algorithm can be drastically simplified using an observation nthistle made: The sides can be counted very similarly to the perimeter just by excluding edges if there is an edge to the "right" of it. That way exactly one edge segment will be counted per side! Very easy to implement and understand.