2024 Day 14 Part 2
This is definitely not what I expected. I was thinking with the moving robots we might be asked to run this for more time and read out the text or something, but instead we need to look for "a Christmas tree". That's honestly super vague and unhelpful, a Chistmas tree could mean many things! I interpreted it as being arranged in a triangle in the grid and only got the answer out of dumb luck! So yeah, honestly I don't like this problem.

(I also burned time thinking about the Chinese Remainder Theorem but that didn't end up applying, all the robots in my input have a cycle of 10403 seconds.)

I guess thinking now (and having glanced at the reddit already) I see some folks talking about looking for rows of robots, maybe you were supposed to track the "connectivity" or "clustering" of the robots? That definitely wasn't clear in the problem text though, and manually looking at 10403 states is not realistic...

Rank 1262. I'm just glad to be done with this mess.