2024 Day 15 Part 1
Another grid problem (I feel like we've had a ton of these this year!) but this time a robot moving boxes. That does add a little bit of trickiness, but not too much. Just need to make sure that the boxes get moved "in reverse" to not obliterate boxes as they get moved.

I did have two bugs though:
1) I initially was skipping moving the robot's location in the grid due to where I computed srcx,srcy. Easy to fix once I noticed.
2) A very dumb bug where I forgot to update my variables keeping track of where the robot was!
These two cost me an annoying amount of time in debugging to figure out...

Rank 444. Looking at my first submission time (with the bug) I was close to leaderboarding in part 1! Lost 6 and a half minutes to the two bugs but only missed the leaderboard by 5 and a half. Oof!