2024 Refactor Day 15
The part 2 code can be used for part 1 as well with a few minor tweaks, so pull that out into a single solve function to be used by both parts. Part 2 just needs an extra step to widen the grid which isn't much at all.

Also, clean up the movement code a little bit.