2024 Rewrite Day 18 Part 2
Someone on reddit had an interesting idea: Do a normal BFS, but keep track of the last good time for each tile. This allows for a pretty quick calculation of when access to the end tile finally breaks thus which falling byte breaks it.

Furthermore, this can be optimized by applying the same idea as dijkstra's: Flood fill from the last accessible tiles before tiles which get blocked sooner. This way less of the space needs to be explored and each tile is only processed once.

All told, this ends up computing the part 2 answer in less time than part 1, though a big part of that is part 1 using a more generic dijkstra's implementation with some extra overhead. It's also not too complex (much simpler than the union-find approach I was originally exploring)!