2024 Day 19 Part 2
Ah, and now we're counting the ways that a design can be constructed. A pretty trivial modification to the part 1 solution.

Rank 655. Jumped up a lot, but still pretty low thanks to the regex sidetrack in part 1. (Even without that there's no way I would have leaderboarded with the times on the leaderboard, most of which are quite unrealistic.)