2024 Day 20 Part 2
Ah, so now we can cheat for up to 20 picoseconds! Thankfully this is not a bad extension of part 1, though I did manage to make some dumb mistakes when enumerating my cheat options. I forgot to account for the varying cheat times (and to account for the new max cheat time) and accidentally treated all cheats as taking 2 picoseconds instead of 20 (or however long they took)!

I'm also annoyed at how long part 2 takes to run. Only a few seconds, but I feel like it should be able to be faster.

Rank 1052. Beyond painful...