2024 Optimize Day 20
Using some clever array slicing it is possible to compute batches of cheats which have moved the same amount. As this moves a bunch of looping down into the C layer it ends up being faster, computing the part 2 solution roughly 20x faster. In fact, it's so efficient that this code can now compute the answer for cheats up to 100 picoseconds in about half a second, nearly 2x as fast as the original solve code!

The key idea is to precompute the "start" and "end" times then shift the arrays based on the "cheat delta" and add them together (plus the cheat time) to get the overall path runtime. Then some vectorized comparisons against the normal solve time and minimum required savings for each cheat delta gets the answer! It takes a little bit more time to think through, but the speedup is well worth it.