2024 Day 22 Part 1
Ah, an RNG problem! This takes me back to 2019 Day 22 which was my favorite problem of that year! The RNG was hidden there but this time it was blatant.

Pretty simple part 1, just correctly implement the RNG and generate the 2000th number in the sequence for each integer in the input.

Rank 249. Missed the leaderboard by 1 minute, 48 seconds. There's no way I would have been able to encode the whole RNG and run it that fast, I was already done within 4 minutes!