2024 Further optimize (and fix) Day 22
Vectorizing part 2 is a little complicated, but feasible. The key idea is to handle all sellers at once when computing the delta sequences and tracking the sell prices per seller/sequence in a giant 5D array. (This is done per seller so that two sellers hitting the same sequence at the same time do not conflict with each other.) Then in a second finalization pass each seller's final sell totals can be added to a global total array (4D for all delta sequences) before getting the final answer as the maximum value in the global totals array.

Also, fix a bug where the initial secret value state was not kept by generate_rngs. According to the problem text in part 2 that should be used in computing the first delta!