2024 Day 23 Part 1
Ah, so we have a bidirectional graph and need to find triplets of fully connected computers. Another simple start, not much to it other than just checking all the possible triplets and counting all the ones that work.

I did have a bug where I was over-counting the triplets initially. (For example, 'a,b,c', 'b,a,c', and 'c,a,b' would all have been treated as distinct!) This was due to me forgetting that computers that were already the "start" of a triplet could appear as connected from other "start" computers! Took me 1 minute and 23 seconds to figure that out, oof!

Rank 396. Even without the silly bug I'd have missed the leaderboard by just under 2 minutes...