2024 Rewrite Day 23
Both parts can make use of a single clique-finding function to find maximal cliques in the graph, the only caveat being that part 1 needs to limit the maximum clique size to 3 which isn't too bad. Additionally, since the new find_largest_cliques function is finding the largest cliques it can discard candidates which cannot reach (or exceed) the current maximum to help optimize the search. (It also makes use of a handled_starts set to avoid duplicate cliques being generated and to help optimize the search.)

Given this new function, the part 1 and 2 solve functions can be drastically simplified and both run in very little time at all.