2024 Optimize Day 23
find_largest_cliques can work faster by progressively pruning down the candidate connection set as each computer is added to the network. The candidate set must be connected to all computers in the group so intersecting the new connection set with the candidate set pre-validates that all candidates are connected to the rest of the group. (This also obviates one of the previous optimizations as this optimization is stronger!)

With this, part 2 runs faster than part 1!