2024 Rewrite Day 24 Part 2
I'm never satisfied until I have a programmatic solution, so here's one for part 2. In some ways it has similar logic to the auto-renaming to detect what gates are what in the binary adder and to find errors. It's definitely a bit messy and there may be smarter/better ways to do this, but it at least works for my input. Not sure if other inputs may have edge cases that I missed here.