2025 Day 3 Part 1
Finding the maximum selection of digits in a sequence, that's a new one! My first pass was dumb because I missed that the order mattered, not sure how I missed that to be honest. But oh well, at least fixing it with nested loops wasn't too bad.

Finished at 2:27. Could have been better, but could have been worse!