2025 Day 3 Part 2
And now it's 12 digits! That definitely ups the ante here. For some dumb reason I started with brute force, I don't know why, and that just ate up time.

After recognizing that that won't work (duh) I finally got to implementing a smart number filter. The key insight here is that Axxxxxxx will always be larger than Bxxxxxxx if A > B, so we just need to pick the largest digit we can that leaves room for picking the other batteries in our bank. It took me an annoying amount of time to notice that (and then coding up the code to do just that) but at least it worked well. (After fixing a dumb typo where I wasn't returning the best value...)

Finally finished at 11:05