2025 Day 5 Part 2
Ah, fun! Counting the total span of the ranges excluding overlaps is a good problem! The only problem is I immediately was thinking of 2021 day 22 which was similar but in 3D space. I had a clever solution there for handling overlaps without splitting the ranges but I couldn't think through how it worked on the fly so it was just a waste of time. Splitting the ranges is much simpler to implement and is what I should have gone with initially.

Finished in 10:59. I definitely should have been faster...