2025 Rewrite Day 6
My initial idea for a simpler parsing was more complicated than I thought, but I realized my grid library has a transpose feature. This makes it much easier to parse the numbers in part 2, and in fact it lets me write a pretty tidy solve function to handle both parts. The only thing that changes is how we grab the number columns after all, and with a tidy parser for the part 2 numbers it fits in well.