2025 Day 7 Part 1
Well, lost some time here by trying to solve the wrong problem. I was trying to count how many unique pathways there are but this problem doesn't care about that, it cares about how many split points get hit in the path down. It's a simple enough modification but it did cost me a couple of minutes in understanding the actual ask of the problem and switching the implementation.

Finished in 4:13