2025 Rewrite Day 7
I was going to clean up and comment the memoized recursive solution but a full iterative rewrite makes the code much simpler and easier to understand so that's a better call. Part 1 just care about the split points that we hit, and part 2 cares about the total timelines which is easy to track in a dictionary. Not much more to it than that!