2025 Day 8 Part 2
Ah, we're trying to create one big circuit! That's not too bad so long as you don't care about being super efficient, just connect one at a time and continually check if you have a completed circuit yet. I'm definitely going to revisit this to optimize the check (maintaining circuit sets will make this much more efficient) but this was quick to code and quick enough to run so I'll take it.

Finished at 16:05