2025 Day 10 Part 1
This is like the light up game! I took an annoying amount of time to parse this and a similarly annoying amount of time to realize why it wasn't terminating (I forgot to break out of the loop when a match was found) so my time here is kind of bad. (It's actually worse than I remember!)

Finished at 15:00