2025 Day 10 Part 2
I clearly don't know how to approach this properly. I got trapped in a graph-based rabbit hole for a very long time while also thinking some linear algebra tricks might help (and I even used some insights from there to try to optimize the search) but never did figure it out. Even worse, it took me forever to think of using z3 to solve it. at least that actually converged, but I clearly need to look at other solutions and learn how to properly solve this.

The code for the graph-based approach is kept here for posterity but will be deleted in the next commit since it's not helpful.

Finished at 1:49:17