2025 Day 11 Part 2
Interesting, we have to find paths that include two specific nodes on the way! I did flail around a bit before realizing that I could compute the steps between the "waypoints" and multiply them together, as well as reducing/segmenting the graph such that computing the steps finishes in a reasonable amount of time. But I got it! (And I was able to use my lib.graph.plot_graph utility in practice to help analyze the input! I'm very glad I finally got that together...)

Also, it looks like there's no possibility of cycles in the input graphs which drastically simplifies the path counting code, I may not have even needed to reduce the graphs thinking about it!

Finished at 36:27