2025 Day 12 Part 1 (There is no Part 2)
So we have something Tetris-adjacent here! I was bracing for a hard packing problem to fit the shapes in but submitted an answer with a super dumb packing test instead (just counting the squares in the pieces and whether they even *possibly* could fit in the shapes) and it was the right answer! I'm guessing it's more likely with larger regions due to how tightly the presents can be packed in practice in bulk but I definitely don't have a hard math proof of this. (I also suspect that certain inputs can be crafted even at larger scales which would break this test but I suspect the problem and inputs are crafted specifically to allow the dumb test to work.) I'll be interested in a more proper solution to be sure, but I'm glad I submitted the stupid answer. I was fully prepared for a 1 minute timeout here!

I left my WIP code for a larger solver largely in-tact in this push just to show the thought process I was going down, I'll clean it up in a follow-up.

We indeed do not have a part 2 on day 12, but my part 2 submission didn't work quite right. I'm going to need to investigate this later and see what's wrong, but for today I just manually clicked the link for part 2. A 4 second delta is fine!

Finished at 23:24 (Part 2: 23:28 because submitting part 2 by script broke)