2025 Rewrite Day 12
Since count-based logic is enough to solve this in practice I can rip out all the extra stuff that was for a general solution. Instead, just check if the presents *must* fit (there's guaranteed to be enough room in the grid) and if not, verify that they *must not* fit (no packing could fit them by pidgeonhole principle). If any regions are left then they need more extensive checks so they should be reported, but inputs appear to be set up such that a true packing does not need to be determined to solve this so this is good enough for a "final" solution (unless I see something on reddit where someone makes an efficient true packing solution).